Refactor SubmitScoreModal to have the submit logic be passed as a prop callback
This makes it easier to test the component
