+++ src/components/SubmitScoreModal/SubmitScoreModal.tsx
@@ -1,12 +1,12 @@
 import { useEffect, useRef, useState } from "react";
 import SubmitScoreModalProps from "./type";
 import SubmissionResponse from "../../types/submissionResponse.type";
-import GameVersion from "src/types/gameVersion.type";
 
 const SubmitScoreModal = ({
   timeScore,
   displayTime,
   gameVersion,
+  submitScore,
   closeModal,
 }: SubmitScoreModalProps) => {
   const nameInputRef = useRef<HTMLInputElement | null>(null);
@@ -18,54 +18,6 @@ const SubmitScoreModal = ({
   useEffect(() => {
     nameInputRef.current!.focus();
   }, []);
-
-  const checkForEmptyName = (name: string) => {
-    if (name === "") {
-      return "Anonymous Trainer";
-    }
-    return name;
-  };
-
-  const checkForEmptyFavoritePokemon = (favoritePokemon: string) => {
-    if (favoritePokemon === "") {
-      return "Missingno";
-    }
-    return favoritePokemon;
-  };
-
-  const submitScore = async (
-    timeInMilliseconds: number,
-    playerName: string,
-    playerFavoritePokemon: string,
-    gameVersion: GameVersion
-  ): Promise<SubmissionResponse> => {
-    try {
-      const data = {
-        name: checkForEmptyName(playerName),
-        score: timeInMilliseconds,
-        favoritePokemon: checkForEmptyFavoritePokemon(playerFavoritePokemon),
-        gameVersion,
-      };
-      const submissionResponse = (await fetch(
-        `${process.env.NEXT_PUBLIC_API_GATEWAY_HTTPS_ENDPOINT}/leaderboard`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            "x-api-key": process.env.NEXT_PUBLIC_X_API_KEY!,
-          },
-          body: JSON.stringify(data),
-        }
-      ).then((data) => data.json())) as SubmissionResponse;
-      return submissionResponse;
-    } catch (e) {
-      console.error("There was an error submitting your score. Try again!");
-      return {
-        success: false,
-        message: ["There is an error with the server. Please try again!"],
-      };
-    }
-  };
 
   const attemptSubmit = async () => {
     const submissionResponse: SubmissionResponse = await submitScore(
